Newly imported audio regions extend project length if needed.
